fix: dual alert messages on approve and deny operations

+++ proto06/panels/src/pages/dashboard/admin-portal/first-inner-pages/second-inner-pages/ea-kpi-questions/local-component/right-side/local-components/main-modals/inner-modals/delete-eval-questions-modal.tsx
@@ -32,12 +32,12 @@ export default function DeactivateEVALQUESTIONSModal(props: DeactivateEVALQUESTI
   }
 
   React.useEffect(()=>{
-    if(EVALQUESTIONSDeactivateState.status === 'succeeded'){
+    if(EVALQUESTIONSDeactivateState.status === 'succeeded' && DeactivateEVALQUESTIONSOpenModal){
       window.alert(`Success: ${EVALQUESTIONSDeactivateState.status?.charAt(0).toUpperCase()}${EVALQUESTIONSDeactivateState.status.slice(1)}`)
       setTimeout(()=>{
         window.location.reload();
       }, 800)
-    } else if(EVALQUESTIONSDeactivateState.status === 'failed') {
+    } else if(EVALQUESTIONSDeactivateState.status === 'failed' && DeactivateEVALQUESTIONSOpenModal) {
       window.alert(`Error: ${EVALQUESTIONSDeactivateState.error}`)
       dispatch(EVALQUESTIONSDeleteActionFailureCleanup());
     }
